feat: add Geometry::point
